+++ frontend/src/components/Tiptap/EditorToolbar.TextSize.tsx
@@ -5,7 +5,7 @@ import { TextSize } from "@edifice-ui/icons";
 import { Dropdown, IconButton, IconButtonProps } from "@edifice-ui/react";
 import { useTranslation } from "react-i18next";
 
-import { useEditorContext } from "./EditorContext";
+import { useEditorContext } from "../../hooks/useEditorContext";
 
 interface Props {
   /**
